added test cases

+++ src/main/java/com/capgemini/constants/BankingConstants.java
@@ -2,7 +2,7 @@ package com.capgemini.constants;
 
 public class BankingConstants {
 
-    public static final int INITIAL_AMOUNT = 500;
+    public static final double INITIAL_AMOUNT = 500;
     public static final String RANDOM_CHARS = "abcdefghvklkhfsty";
     public static final int RANDOM_STRING_LENGTH = 6;
 }
